requirements.txt
Pillow==10.0.1
Pillow==10.0.1
rembg==2.0.50
streamlit==1.27.1
streamlit_image_comparison==0.0.4
~treamlit==1.24.1
